Areas de cobertura das UBS da cidade de Sao Paulo

O dataset foi obtido da Secretaria Municipal de Saude por meio de pedido de acesso a informacao. Traz os polígnos dos territórios das Unidades Basicas de Saude, identificadas pelo número CNES de cada equipamento. Foram adicionadas variaveis com a quantidade de medicos e enfermeiros que trabalham em cada unidade.

ubs_malha

Format

Um data frame com 94765 linhas e 10 colunas

cnes

Codigo de identificacao da UBS

nomeubs

Nome da UBS

sts

Superintendencia de saude responsavel pela unidade

crs

Coordenadoria regional de saude responsavel pela unidade

subpref

Prefeitura regional onde a UBS esta localizada

geometry

Geometria dos polígonos